Applied code changes to make the library work better with 64bit systems.
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@22 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/angelscript/source/as_callfunc.cpp
@@ -57,14 +57,14 @@ int DetectCallingConvention(bool isMethod, const asUPtr &ptr, int callConv, asSS
 	int base = callConv;
 	if( !isMethod )
 	{
-		if( base == asCALL_GENERIC )
+		if( base == (int)asCALL_GENERIC )
 			internal->callConv = ICC_GENERIC_FUNC;
 		else
 			return asNOT_SUPPORTED;
 	}
 	else
 	{
-		if( base == asCALL_GENERIC )
+		if( base == (int)asCALL_GENERIC )
 			internal->callConv = ICC_GENERIC_METHOD;
 		else
 			return asNOT_SUPPORTED;
